python3/sklearn/emotion/requirements.txt
sklearn
mecab-python3
